fix: bundleMDX with custom options
