+++ package.json
@@ -21,7 +21,7 @@
 		"@notionhq/client": "^0.4.0",
 		"@supabase/supabase-js": "^1.23.0",
 		"axios": "^0.21.4",
-		"esbuild": "^0.13.2",
+		"esbuild": "^0.12.19",
 		"feed": "^4.2.2",
 		"framer-motion": "^4.1.17",
 		"gray-matter": "^4.0.3",
