+++ src/components/mdx/index.tsx
@@ -1,9 +1,4 @@
-import { promises as fs } from "fs";
-import path from "path";
-
 import { MDXProvider } from "@mdx-js/react";
-import { bundleMDX } from "mdx-bundler";
-import Image from "next/image";
 import Link from "next/link";
 import React, {
 	createElement,
@@ -13,49 +8,12 @@ import React, {
 	ReactNode,
 } from "react";
 import { FiLink } from "react-icons/fi";
-import remarkSlug from "remark-slug";
 import styled from "styled-components";
 
 import { MDXCodeBlock } from "components/mdx/code";
 import { LinkedHeaderIconWrapper } from "styles/blog";
 import { Paragraph } from "styles/typography";
 import { useHover } from "utils/hooks";
-
-if (process.platform === "win32") {
-	process.env.ESBUILD_BINARY_PATH = path.join(
-		process.cwd(),
-		"node_modules",
-		"esbuild",
-		"esbuild.exe"
-	);
-} else {
-	process.env.ESBUILD_BINARY_PATH = path.join(
-		process.cwd(),
-		"node_modules",
-		"esbuild",
-		"bin",
-		"esbuild"
-	);
-}
-
-export async function bundleMDXWithOptions(filename: string) {
-	const mdxSource = await fs.readFile(filename, "utf8");
-
-	return bundleMDX(mdxSource, {
-		cwd: path.dirname(filename),
-		xdmOptions(options) {
-			options.remarkPlugins = [...(options.remarkPlugins ?? []), remarkSlug];
-			options.rehypePlugins = [...(options.rehypePlugins ?? [])];
-
-			return options;
-		},
-		esbuildOptions(options) {
-			options.platform = "node";
-
-			return options;
-		},
-	});
-}
 
 export const ImageWrapper = ({ alt, src }: { alt: string; src: string }) => {
 	const type = src.slice(-3);
@@ -79,14 +37,14 @@ export const ImageWrapper = ({ alt, src }: { alt: string; src: string }) => {
 		);
 	}
 	return (
-		<Image
+		<img
 			{...{ alt, src }}
 			loading="lazy"
-			// style={{
-			// 	maxWidth: "var(--max-width)",
-			// 	width: "100%",
-			// 	borderRadius: "var(--border-radius)",
-			// }}
+			style={{
+				maxWidth: "var(--max-width)",
+				width: "100%",
+				borderRadius: "var(--border-radius)",
+			}}
 		/>
 	);
 };
